Add a 5-second tutorial & reorganize main info.
This will finally provide an authoritative source for the most trivial
rack use cases.

I also changed the site to a basic two-column layout, with the
incredibly-long News section on the left and everything else on the
right. This makes it much easier to scroll down and see what information
is available, as opposed to getting lost in the News and Releases lists.

The CSS has been tested in Chrome and Safari.